refactor: split training utils
